Fix failing dashboard spec

+++ cypress-tests/cypress.config.js
@@ -50,13 +50,19 @@ module.exports = defineConfig({
       on("task", {
         deleteFolder(folderName) {
           return new Promise((resolve, reject) => {
-            rmdir(folderName, { maxRetries: 10, recursive: true }, (err) => {
-              if (err) {
-                console.error(err);
-                return reject(err);
-              }
-              resolve(null);
-            });
+            if (fs.existsSync(folderName)){
+              rmdir(folderName, { maxRetries: 10, recursive: true }, (err) => {
+                if (err) {
+                  console.error(err);
+                  return reject(err);
+                }
+                return resolve(null);
+              });
+            }
+            else{
+              return resolve(null);
+            }
+ 
           });
         },
       });
